automate_observation: Added workaround for pixel storage bug.
A storage bug for devices shipping with Android 11 has caused Hubble
results to be inaccessible to adb shell.

We thus work around that using `adb backup`. This unfortunately
requires manual intervention, but is the only way to do it currently
without modifying Hubble.

Also added a backup method in syscall_wrapper.py to accommodate this.