Using worker thread to avoid ANR
Hubble uses the Ui thread to do all the tasks (get* and write*).
If it takes a longer time (more than 5 sec), we may get an ANR
error. It's better to use another thread to do these tasks and
we could also have some status to show on MainActivity. This
change also bumps up Hubble's version to "1.1.0".