Add support for new permissions introduced in Android 12
This commit updates the Permission Tester and related apps
with support for Android 12 (API level 31); the BasePermissionTester
extensions have been updated to include tests for the new permissions
introduced in Android 12. This Android release also included support
for a new internal permission protection level; permissions declared
with this protection level are no longer granted to apps signed with
the platform's signing key, but instead a requesting app must satisfy
one of the other permission protection flags. This commit includes a
new instrumentation test that adopts the shell identity to test the
new internal protection permissions.