uraniborg/Hubble: Upgrade Hubble to v.1.2.0.
Minor updates and changes to get Hubble to target API/SDK level 33.
Oficially, this updates the versionName to "1.2.0" and versionCode
to 4.

AndroidManifest.xml:
MainActivity has to be explicitly exported due to change in Android 12.
The necessary permission for Hubble to successfully probe packages is
also added, i.e. QUERY_ALL_PACKAGES.

MainActivity.java:
Version const is updated to reflect the versionName.